refactor: extracted more stuff to traits
