+++ src/HeaderToken.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Staffbase\plugins\sdk;
+
+use Staffbase\plugins\sdk\SSOData\HeaderSSOData;
+
+class HeaderToken
+{
+    use HeaderSSOData, SSOTokenTrait;
+}
