+++ src/SSOData/HeaderSSOClaims.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Staffbase\plugins\sdk\SSOData;
+
+trait HeaderSSOClaims
+{
+    public static string $CLAIM_BRANCH_ID = "branchId";
+    public static string $CLAIM_USER_ID = "userId";
+    public static string $CLAIM_TOKEN_ID = "tokenId";
+}
